a lot
I updated all code to match my new writing style for c++.

I added more macros to event listener to make checking if the event matches and that the data is correct and valid, I then implemented this into code.

The logic for the events used in ai I moved to c++ so that it's duplicated for the player, the player won't have a death screen they'll just respawn similar to ai because it's easier.

I added what I hope should be a working spawn/respawn system, but we will see when I test tomorrow