fixed the death/respawn system and fixed events stuff
basically it was the GetOwner() being called in character that was causing the issue so now it correctly does GetWorld() for actors and GetOwner()->GetWorld() for actor components.

fixed some formatting issues with some code.

ai controller now checks if the blackboard is valid in the seen enemy set as well on the set seen enemy